move split.py contents to scripts/split.py
