+++ pyproject.toml
@@ -14,7 +14,6 @@ requires-python = ">=3.8"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
 ]
 dynamic = ["dependencies"]
