+++ src/splat/__init__.py
@@ -8,3 +8,5 @@ from . import util as util
 from . import disassembler as disassembler
 from . import platforms as platforms
 from . import segtypes as segtypes
+
+from . import scripts as scripts
